boyscout some of the network code
