+++ signer/src/network/libp2p/mod.rs
@@ -7,9 +7,11 @@ use libp2p::gossipsub::IdentTopic;
 mod errors;
 mod event_loop;
 mod swarm;
+mod network;
 
 pub use self::errors::SignerSwarmError;
 pub use self::swarm::SignerSwarmBuilder;
+pub use self::network::P2PNetwork;
 
 /// The default port for the libp2p network
 pub const DEFAULT_P2P_PORT: u16 = 4122;
